+++ src/pages/List/List.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../../components/Header";
 import DonationsList from "./components/DonationList";
-import ThisMonthsChart from "./components/ThisMonthChart/ThisMonthsChart";
+import ThisMonthsChart from "./components/ThisMonthsChart/ThisMonthsChart";
 import MyCredit from "./components/MyCredit/MyCredit";
 import CreditContextProvider from "../../components/CreditContextProvider";
